+++ scripts/copy-files.js
@@ -58,7 +58,7 @@ function createPackageFile() {
     } = packageData;
 
     const minimalPackage = {
-      name: 'material-ui',
+      name: '@logicline/material-ui',
       author,
       version,
       description,